Fix support of multiple instances of the Router for the same topology
The EventEmitter backend was bugged because the same handler was used
between multiple BusClient in the same process.

In order to fix the bug, it's necessary to have a key and this one is
available on connect or start only. Then the EventEmitter is created
only where appropriate and not in the constructor.

Because events can be registered before starting, a pending list is used
in order to subscribe to the event on start or connect.