Add a cache based on id provided in the topics
The ids are something like
  text[@text...]@guid

The cache has a key for each id where the regexes are grouped by key. It
prevents running regex.test() on all subscriptions. Only a subset is
tested. For subscriptions without ids, then a special key '_' is used
for this subset. Theses regexes are always checked with all topics.

Usual check with topic (with id) is:
  1. '_'
  2. 'text[@text...]@guid'
or:
  1. '_'
  2. 'orcname@guid'

It's possible to have several id with one topic, for example:
  orcname@guid::text[@text...]@guid

In thise case, the topics are checked against (in this order)
  1. '_'
  2. 'text[@text...]@guid'
  3. 'orcname@guid'