Switch finished|error events according to the transport
It depends of the transport used by the command. If the command comes
from axon, then the finished|error events are sent only to this
transport.

Note that for the horde is very different. It should be routed according
to the node and not just by the transport. It's not implemented..