Improve significantly the cache by extracting the simple UUID too
It's very efficient in the case of all subscriptions that rely on Xcraft
bus message id. It's the case for most messages :-)

This commit prevent very large list of test of regex (like 10000 or
event more). Now most of times it's fine with less than 10 regex tests in
the worse cases.

I think that I can improve it more...