Save the socket in the ARP table
Command sent from Portal:

Portal:   orcName token == token && socket == null : event emitter
Thrall:   orcName token == token && socket != null : axon (direct parent)
Business: orcName token != token && socket != null : axon (indirect parent)

  Event Emitter
      ,-.
      | v    Axon           Axon
     Portal ------> Thrall ------> Business

In the context ofr Thrall, the token is the real origin, it's the
Portal's token. But in the case of Business, the token can not be the
Portal's token because Thrall is in the middle. The token is in this
case the Thrall's token.