Rework the whole lines stuff
The lines are still connected via a subscribe but the tables are updated
when reveiving an event forwarded by the warehouse (the deeper node).
All nodes will receive this event in order to maintain their own lines
table.

The main goal of this rework is for supporting bidirectional links for
commands between nodes.