Handle cert rejection from our side because BoringSSL is boring
Without this trick, it's not possible to detect when a client is
rejected because the client certificate is not authorized. With nodejs
it works because an exception is received on the client side. But with
Electron, nothing is received as error. The socket is just closed
without explicit reason.

With this change, we break the connection with an explicit new 'reject'
event from Axon.